refactor(web): one component per file in the sidebar
AppSidebar held five components. Split them along their usage:
Wordmark and SidebarBody under AppSidebar, SidebarNav (with its nav
constants) and OrganizationMenu under SidebarBody.
