+++ apps/web/src/app/(app)/components/DashboardShell/components/AppSidebar/components/SidebarBody/components/OrganizationMenu/index.ts
@@ -1,0 +1,1 @@
+export { OrganizationMenu } from "./OrganizationMenu";
