+++ apps/web/src/app/(app)/components/DashboardShell/components/AppSidebar/components/SidebarBody/components/SidebarNav/index.ts
@@ -1,0 +1,1 @@
+export { SidebarNav } from "./SidebarNav";
