+++ apps/web/src/app/(app)/components/DashboardShell/components/AppSidebar/components/SidebarBody/index.ts
@@ -1,0 +1,1 @@
+export { SidebarBody } from "./SidebarBody";
