+++ apps/web/src/app/(app)/components/DashboardShell/components/AppSidebar/components/Wordmark/index.ts
@@ -1,0 +1,1 @@
+export { Wordmark } from "./Wordmark";
